Add files via upload
We have answered these 5 questions through our data analysis mainly using pandas and matplotlib library.

Q1. What is the overall sales trend?
Q2. Which are the Top 10 products by sales?
Q3. Which are the Most Selling Products?
Q4. Which is the most preferred Ship Mode?
Q5. Which are the Most Profitable Category and Sub-Category?